ref: remove sentry.utils.functional
many of the functions were only used once so I moved them to their consuming modules
