+++ src/sentry/api/serializers/models/activity.py
@@ -1,5 +1,3 @@
-import functools
-
 from sentry.api.serializers import Serializer, register, serialize
 from sentry.api.serializers.models.commit import CommitWithReleaseSerializer
 from sentry.models.activity import Activity
@@ -9,7 +7,6 @@ from sentry.models.pullrequest import PullRequest
 from sentry.services.hybrid_cloud.user.serial import serialize_generic_user
 from sentry.services.hybrid_cloud.user.service import user_service
 from sentry.types.activity import ActivityType
-from sentry.utils.functional import apply_values
 
 
 @register(Activity)
@@ -65,9 +62,9 @@ class ActivitySerializer(Serializer):
         else:
             pull_requests = {}
 
-        groups = apply_values(
-            functools.partial(serialize, user=user),
-            Group.objects.in_bulk(
+        groups = {
+            k: serialize(v, user=user)
+            for k, v in Group.objects.in_bulk(
                 {
                     i.data["source_id"]
                     for i in item_list
@@ -78,8 +75,8 @@ class ActivitySerializer(Serializer):
                     for i in item_list
                     if i.type == ActivityType.UNMERGE_SOURCE.value
                 }
-            ),
-        )
+            ).items()
+        }
 
         return {
             item: {
